Code Clean up and code file split.
