+++ webstack/libs/frontend/src/lib/ui/components/party/components/index.ts
@@ -6,4 +6,6 @@
  * the file LICENSE, distributed as part of this software.
  */
 
-export * from './Party';
+export * from './PartyHub';
+export * from './PartyInstance';
+export * from './PartyStore';
